Install and configure i3
Install and configure the i3 window manager, which
seems like a great candidate for a Linux
window manager.

The method for installing i3 as the default window manager
may not work if using this script with Debian (I'm
not sure if Debian uses `lxdm`). Hopefully a similar
strategy of editing configuration files will work.