Configure xinput mouse on i3 start
Add a utility script for configuring my mouse to
use reverse scrolling and be less sensitive. Set up
i3 to run this script on boot.

Using i3 appears to be the best method for running this script.
I also tried `~/.profile`, a `reboot` cron, and `~/.xinitrc`,
and none of them appeared to actually call the script when
I wanted.

I left the option of specifying a `~/.profile.local` in case
I decide that could be useful in the future.