Productionize bash scripts
Modify my bash scripts to make them more robust - specifically, set
the PATH to a restrictive a value as possible, so any PATH tampering
will have no impact on the script (this step may not be
necessary/possible if a program relies on path modifications).

Additionally, refactor the process for displaying incorrect script
usage.